Luke: Add support for spanning resources
- span_resources in IMS will iterate through all known resource types
  and create independent index 'types' in ElasticSearch based on known
resources.
- _map_resources will iterate through the resource schemas and create a
  best-guess appropriate mapping.  More features will follow-

TODO: Need to discuss further support for good indexing types such as a
date type in the resource which we can search and index with, as well as
a concrete definitionf or our geo-points.  Complex types will need a
javascript filter and will be slow to initially index (not too much of a
problem).